p2 priority pass

+++ examples/file_sd.be
@@ -1,10 +1,9 @@
 import os
 
-var path = "/P2FILE.TXT"
+var path = "/P2FEXMP.TXT"
 
-try
-    os.remove(path)
-except .. as e, m
+if os.path.exists(path)
+    raise "io_error", path + " already exists"
 end
 
 var f = open(path, "w")
@@ -15,3 +14,4 @@ print("exists:", os.path.exists(path))
 print("text:", open(path, "r").read())
 print("bytes:", open(path, "r").readbytes().tohex())
 print("remove:", os.remove(path))
+print("file sd example done")
